Restructure question generator file system
Own file for each topic rather than own file for each part of a question object.
