+++ src/logic/questionGenerators/fractionsGenerator.js
@@ -1,0 +1,8 @@
+// Exports helpers for when questionGenerator has selected the "fractions" topic
+
+import { randomInt, chooseRandomFromArray } from "../lowLevelHelpers";
+
+
+function getFractionsQuestion(difficulty) {
+
+}
